Add derived attributes and tidy
The four derived attributes are added and the init code is generally
tidied, including removing a duplicate race statement, grouping similar
statements such as level and adding comments.

In particular, the logic around the level cap and level margin of error
is collected into a single pass for neatness. .copy() is also added to
the dictionaries to avoid the alias trap.