Add override method to import existing characters
Rather than recreating the whole levelling up history of a character
with the increase_skill and level_up methods, this method can be used to
recreate the character as they are after a level up, while still
benefitting from the calculations derived from the character init. The
validity of the user input is not checked - this will be dealt with
later.

Also save the birthsign to avoid having to input it again on the
override method. Otherwise, it still only works with the apprentice
sign.