Split showStatus into journal and minmax
showStatus was a placeholder method designed to print "everything else"
other than the progressToLevelUp info. I realised it made more sense to
separate them out into the journal and "hidden"/derived minmax info,
which plots how much of the budget of "spare" skill increases has been
used while still leaving room to efficiently increase an attribute to
100.

It took me a little while to figure out how the skills were sorted in
the journal but I got there in the end. Rather than hardcoding
character_journal_skill_order it seemed neater to calculate it on
startup, as performance is not an issue for a small codebase like this.