Add training tracker, times_trained_this_level
While not strictly necessary for min-maxing, training is helpful and
only available in limited quantities. Tracking times trained in a level
can be helpful towards the end of a level when planning which skills to
round out the level with and whether to train those.

All that is tracked is the times trained in a level, which is reset when
levelling up or reverting to the last level. Minmax calculations are
unchanged. The times trained is plotted in the level up progress chart.