Remove skills with attribute 100 from outputs
Once an attribute has reached 100 it's irrelevant how many skill
increases have happened in the skills governed by that attribute, so
such attributes are removed from the progress to level up and minmax
outputs. Increases of major skills of such attributes are of course
still tracked as before.

Note that the underlying data is still recorded in the
level_up_attribute_bonuses and wasted_skill_ups, should it be wanted for
whatever reason.

Also adds a check in minmax to ensure that when all attributes are 100
the current level and health match the ideal maximums.

As part of the refactor the .format method is replaced with the
generally superior f-strings and repetitive long strings are replaced
with loops.