Improve level_up and rename to levelUp
The level_up method had a number of deficiencies, mainly revolving
around a naivety of the limit of 100 for attributes. These are addressed
here.

Rather than blindly increasing three attributes every time, we check
which attributes are under 100, and can therefore be increased. If there
are none, we raise an error. If there are 1 or 2 available we increase
that many, otherwise we increase 3 as before.

Autodetection will proceed with no input if there is only one possible
choice of attributes (i.e. number_of_attributes_to_raise == available
attributes to raise).

On successful autodetection with 4 or more attributes to choose from, we
prompt before continuing in case the wrong attributes have been
detected. This is basically to catch the case where three +5s are
available but luck needs increasing.

With manual attribute input, we now compare against under_100_attributes
instead of all_attributes and check the length of the input against
number_of_attributes_to_raise instead of 3.

When actually raising attributes, guards against overshooting 100 are
now in place.

Finally, the slightly vague variable "attributes" is renamed
attributes_to_raise for clarity.