Fix spelling and finish placeholder validations
Correct instances of 'hand to hand' in the race section to match
all_skills spelling 'hand-to-hand'.

The __init__ methods on the two classes had placeholder printings of 'uh
oh', which isn't actually useful and allows obviously broken classes to
be instantiated. We replace these with exceptions with useful (but
terse) output in the same fashion as the later methods.