Finish birthsign support, refactor derived things
The rest of the birthsigns are pretty straightforward so we add them
here. However, to avoid further repetitious code we refactor magicka
calculations (along with fatigue and encumbrance) to be less "stateful",
although health can't be (easily) included due to its reliance on the
history of the encumbrance attribute.

The journal now includes the birthsign. We also update the README to
reflect the changes.

Finally, to port saved characters to the new format
(magicka_racial_bonus and magicka_birthsign_bonus present), the
rebuildCharacter function is added. This probably forms the basis of a
future improved save system which properly serialises characters.