Add skillLevels method
It became apparent that it would be useful to have a birds-eye view of
how skills are progressing, and so this method is added. The ability to
filter by specialisation is available from the outset because I only
really care about mastering magical skills, and I imagine the same
probably applies to combat and stealth characters.

As always, we update the README to reflect the new feature. People may
not realise that 'magic' is one of any specialisations, though, which
underscores the need for proper documentation.