Refactor increase_skill to output new skill level
While adding the output, I noticed that there was duplicated code across
the if/else statement. Since the if block is essentially optional code
for major skill increases, we remove the duplicate lines from the major
skill code and turn the else code into "universal" code for both (major
+ minor) cases.

We also add a few comments to make the outline of the function clear.