Fix reintroduced bug in increase_skill
Testing the parent commit's code revealed the reason I duplicated the
code across the if/else - the level_up_attribute_bonuses can only be
properly converted after the final major skill is increased. I remember
having this realisation before with the same bug, so there's a lesson to
be learned about commenting traps like this.

Anyway, rather than just reverting the code we restructure it a little
to keep the nice self-contained if statement. The if/else arose to avoid
testing for having over 10 major skill increases. We let go of this and
put the test with the others, as we prioritise legibility over
performance for this repo. Always running the test allows us to shunt
the major skill code to after the universal code, as there is no longer
a risk of increasing the skill when we shouldn't.

Incidentally, we add a "quiet" option to the method as, while confirming
the new code works, I noticed the rebuildCharacter function calls the
increase_skill method potentially quite a lot, which might be
irritating.