Update save system to JSON
As suggested in 5182486c9f60aac4fb600505ccdd8344a980a99a, we expand on
the idea of the (now obsolete) rebuildCharacter method to rewrite the
save/load system to be based on JSON data instead of pickling.

There's a subset of data that can be used to recreate a character
copmletely, which we extract into a dictionary and save as JSON. We
pretty-print because performance isn't foreseeably an issue and it makes
it easier to read save files.

The new loadCharacter largely mimics the old rebuildCharacter, with some
changes:

- the input is a file, read into a dictionary from the JSON. Because
  integers aren't allowed as names (~dictionary keys) we add a step to
  recover the normal format
- we recover the character class object by checking if the details match
  a default class and building the custom class if not
- the rebuilding algorithm set the wrong skills in the override method
  and compensated with an unnecessary line later. Now we just set the
  skills right from the beginning
- we remove the check for increasing major skills by 10 because it won't
  even catch cases where multiple major skills are collectively
  increased by 10 and it's easier to just let a save error out anyway.
  Who's going to be importing invalid characters anyway? This is a
  passion project ffs
- we remove the pointless 'Something went wrong' catch-all except and
  add a more meaningful FileNotFoundError catch
- we return the character instead of messing around with globals

Also update gitignore to reflect new default save name and update the
loadCharacter call by wrapping it in the __name__ stuff.