Stylistic changes
Now that characters are rebuilt on load we can easily change cosmetic
stuff like this. There was a mix of cases for method names so we
standardise on camelCase for methods and snake_case for other
attributes.

We also change "== None" to "is None" as per PEP 8. I'm not willing to
fully comply with that though, this is good enough for now.