Add freeSkills method
When I hadn't trained in a level and was about to level up it was
annoying trawling through skills to see which I could just train 5
times, hence this method.

Major skills are less useful but are still handy if you've already hit
two +5 attribute bonuses and just want to spam a major skill to level
up.